updated to version 3.04
git-svn-id: https://subversion.assembla.com/svn/impresscmsaddons/modules/iforum/trunk@6075 7a4d4d67-0c3a-0410-b24c-89d50c169223

+++ sql/upgrade_230.sql
@@ -5,7 +5,7 @@ CREATE TABLE `bb_moderates` (
   `mod_desc` varchar(255) NOT NULL default '',
   `uid` int(10) NOT NULL default '0',
   `ip` varchar(32) NOT NULL default '',
-  `forum_id` int(4) NOT NULL default '0',
+  `forum_id` smallint(4) NOT NULL default '0',
   PRIMARY KEY  (`mod_id`),
   KEY `uid` (`uid`)
 ) TYPE=MyISAM;
